requirements.txt
pytesseract
Pillow
streamlit
